third_party/WebKit/LayoutTests/virtual/user-activation-v2/fullscreen/full-screen-iframe-zIndex-expected.txt

CONSOLE WARNING: line 20: Failed to execute 'requestFullscreen' on 'Element': API can only be initiated by a user gesture.
This tests that an element with a positive z-index appears behind the full screen iframe element. After entering full screen mode, the whole screen should be white. Click go full screen to run the test.

EVENT(webkitfullscreenerror)
END OF TEST
